requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
pypdf2>=3.0.1
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
google-generativeai>=0.3.2
